refactor: split more accounts services into files
