+++ src/controllers/accounts.controller.ts
@@ -5,8 +5,7 @@ import {
   endpointsTypes,
 } from 'shared-types';
 import { CustomResponse } from '@common/types';
-import * as accountsService from '@services/accounts.service';
-import * as accountsServiceNew from '@services/accounts';
+import * as accountsService from '@services/accounts';
 import { removeUndefinedKeys } from '@js/helpers';
 import Accounts from '@models/Accounts.model';
 import { ValidationError, NotFoundError, Unauthorized } from '@js/errors';
@@ -16,7 +15,7 @@ export const getAccounts = async (req, res: CustomResponse) => {
   const { id: userId } = req.user;
 
   try {
-    const accounts = await accountsServiceNew.getAccounts({ userId });
+    const accounts = await accountsService.getAccounts({ userId });
 
     return res.status(200).json({
       status: API_RESPONSE_STATUS.success,
@@ -32,7 +31,7 @@ export const getAccountById = async (req, res: CustomResponse) => {
   const { id: userId } = req.user;
 
   try {
-    const account = await accountsServiceNew.getAccountById({ userId, id });
+    const account = await accountsService.getAccountById({ userId, id });
 
     return res.status(200).json({
       status: API_RESPONSE_STATUS.success,
@@ -61,7 +60,7 @@ export const createAccount = async (req, res) => {
       });
     }
 
-    const account = await accountsServiceNew.createAccount({
+    const account = await accountsService.createAccount({
       accountCategory,
       currencyId,
       name,
