+++ src/services/accounts/index.ts
@@ -1,2 +1,5 @@
 export * from './get-accounts';
 export * from './create-account';
+export * from './delete-account';
+export * from './update-account';
+export * from './update-account-balance';
