+++ src/services/accounts/update-account-balance.ts
@@ -1,0 +1,120 @@
+import { TRANSACTION_TYPES } from 'shared-types';
+import * as Accounts from '@models/Accounts.model';
+import { withTransaction } from '@services/common';
+
+import { getAccountById } from '@services/accounts';
+
+const calculateNewBalance = (amount: number, previousAmount: number, currentBalance: number) => {
+  if (amount > previousAmount) {
+    return currentBalance + (amount - previousAmount);
+  } else if (amount < previousAmount) {
+    return currentBalance - (previousAmount - amount);
+  }
+
+  return currentBalance;
+};
+
+const defineCorrectAmountFromTxType = (amount: number, transactionType: TRANSACTION_TYPES) => {
+  return transactionType === TRANSACTION_TYPES.income ? amount : amount * -1;
+};
+
+// interface updateAccountBalanceRequiredFields {
+//   accountId: number;
+//   userId: number;
+//   transactionType: TRANSACTION_TYPES;
+//   currencyId: number;
+// }
+
+// At least one of pair (amount + refAmount) OR (prevAmount + prefRefAmount) should be passed
+// It is NOT allowed to pass 1 or 3 amount-related arguments
+
+/** For **CREATED** transactions. When only (amount + refAmount) passed */
+// export async function updateAccountBalanceForChangedTxImpl(
+//   {
+//     accountId,
+//     userId,
+//     transactionType,
+//     amount,
+//     refAmount,
+//     currencyId,
+//   }: updateAccountBalanceRequiredFields & { amount: number; refAmount: number },
+// ): Promise<void>;
+
+// /** For **DELETED** transactions. When only (prevAmount + prefRefAmount) passed */
+// export async function updateAccountBalanceForChangedTxImpl({
+//   accountId,
+//   userId,
+//   transactionType,
+//   prevAmount,
+//   prevRefAmount,
+//   currencyId,
+// }: updateAccountBalanceRequiredFields & {
+//   prevAmount: number;
+//   prevRefAmount: number;
+// }): Promise<void>;
+
+// /** For **UPDATED** transactions. When both pairs passed */
+// export async function updateAccountBalanceForChangedTxImpl({
+//   accountId,
+//   userId,
+//   transactionType,
+//   amount,
+//   prevAmount,
+//   refAmount,
+//   prevRefAmount,
+//   currencyId,
+//   prevTransactionType,
+// }: updateAccountBalanceRequiredFields & {
+//   amount: number;
+//   prevAmount: number;
+//   refAmount: number;
+//   prevRefAmount: number;
+//   prevTransactionType: TRANSACTION_TYPES;
+// }): Promise<void>;
+
+export async function updateAccountBalanceForChangedTxImpl({
+  accountId,
+  userId,
+  transactionType,
+  amount = 0,
+  prevAmount = 0,
+  refAmount = 0,
+  prevRefAmount = 0,
+  prevTransactionType = transactionType,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+}: any): Promise<void> {
+  const account = await getAccountById({ id: accountId, userId });
+
+  if (!account) return undefined;
+
+  const { currentBalance, refCurrentBalance } = account;
+
+  const newAmount = defineCorrectAmountFromTxType(amount, transactionType);
+  const oldAmount = defineCorrectAmountFromTxType(prevAmount, prevTransactionType);
+  const newRefAmount = defineCorrectAmountFromTxType(refAmount, transactionType);
+  const oldRefAmount = defineCorrectAmountFromTxType(prevRefAmount, prevTransactionType);
+
+  // TODO: for now keep that deadcode, cause it doesn't really work. But when have time, recheck it past neednes
+  // if (currencyId !== accountCurrencyId) {
+  //   const { rate } = await userExchangeRateService.getExchangeRate({
+  //     userId,
+  //     baseId: currencyId,
+  //     quoteId: accountCurrencyId,
+  //   }, { transaction });
+
+  //   newAmount = defineCorrectAmountFromTxType(amount * rate, transactionType)
+  // }
+
+  await Accounts.updateAccountById({
+    id: accountId,
+    userId,
+    currentBalance: calculateNewBalance(newAmount, oldAmount, currentBalance),
+    refCurrentBalance: calculateNewBalance(newRefAmount, oldRefAmount, refCurrentBalance),
+  });
+
+  // TODO: call Balances service to update balances
+}
+
+export const updateAccountBalanceForChangedTx = withTransaction(
+  updateAccountBalanceForChangedTxImpl,
+);
